Create some common form components
Certain information will be reused in other intake forms
