+++ frontend/src/components/note/NoteCard.vue
@@ -134,15 +134,6 @@ p {
   color: #38598a;
 }
 
-.p-card {
-  border-style: solid;
-  border-width: 1px;
-
-  :deep(.p-card-content) {
-    padding-bottom: 0;
-  }
-}
-
 .note-content {
   white-space: pre;
   text-wrap: balance;
