+++ frontend/src/locales/en-CA.json
@@ -2,6 +2,10 @@
   "cancelButton": {
     "btnText": "Cancel"
   },
+  "contactCard": {
+    "contactTooltip": "To update your contact details, click on your username in the top right corner of the page and select “contact profile.”",
+    "header": "Who is the primary contact regarding this project?"
+  },
   "contactHistoryList": {
     "noItems": "No items found."
   },
@@ -138,6 +142,20 @@
     "linkHousingSubmissions": "My drafts and previous entries.",
     "linkSubmissions": "Back to Housing"
   },
+  "locationCard": {
+    "acceptableCoordinates": "The accepted coordinates are to be decimal degrees (dd.dddd) and to the extent of the province.",
+    "addressSearchPlaceholder": "Type to search the address of your housing project",
+    "addressTooltip": "A civic address contains a street name and number where a non-civid does not.",
+    "header": "Provide one of the following project locations",
+    "latitudePlaceholder": "Latitude",
+    "localityPlaceholder": "Locality",
+    "longitudePlaceholder": "Longitude",
+    "provideLatitude": "Provide a coordinate between 48 and 60",
+    "provideLongitude": "Provide a coordinate between -114 and -139",
+    "provincePlaceholder": "Province",
+    "showOnMap": "Show on map",
+    "streetAddressPlaceholder": "Street address"
+  },
   "loginButton": {
     "login": "Log in",
     "logout": "Log out"
@@ -211,28 +229,25 @@
   },
   "submissionIntakeForm": {
     "additionalInfoTooltip": "Provide us with additional information - short description about the project, project website link, or upload a document.",
-    "addressTooltip": "A civic address contains a street name and number where a non-civid does not.",
+    "additionalLocationCard": "Provide additional location details (optional)",
     "appliedPermitsShareNotification": "Sharing this information will authorize the navigators to seek additional information about this permit.",
     "appliedPermitsTooltip": "Early information on your permitting needs will help us coordinate and expedite the authorization process.",
     "bcHousingTooltip": "BC Housing welcomes the opportunity to work with individuals and organizations to create affordable housing solutions.",
-    "contactTooltip": "To update your contact details, click on your username in the top right corner of the page and select “contact profile.”",
+    "financiallySupportedCard": "Is this project associated with any of the following?",
+    "hasRentalUnitsCard": "Will this project include rental units?",
+    "investigatePermitsCard": "Select all provincially issued permits you think you might need (optional)",
     "isRegisteredTooltip": "Are you registered with OrgBook BC?",
     "multiFamilyTooltip": "Multi-family dwelling: a residential building that contains two or more attached dwellings, including duplex, triplex, fourplex, townhouse, row houses, and apartment forms.",
-    "potentialPermitsTooltip": "Early information on your permitting needs will help us coordinate and expedite the authorization process.",
-    "rentalUnitsTooltip": "Rental refers to a purpose built residentual unit, property, or dwelling that is available for long term rent by tenants.",
-    "contactCard": "Who is the primary contact regarding this project?",
-    "projectApplicantTypeCard": "Are you applying as an individual or as a business?",
-    "projectNameCard": "Tell us your project name",
-    "singleFamilySelectedCard": "Select the types of residential units being developed",
-    "hasRentalUnitsCard": "Will this project include rental units?",
-    "financiallySupportedCard": "Is this project associated with any of the following?",
-    "projectDescriptionCard": "Help us learn more about your housing project",
     "naturalDisasterCard": "Has the location of this project been affected by natural disaster?",
-    "projectLocationCard": "Provide one of the following project locations",
-    "additionalLocationCard": "Provide additional location details (optional)",
+    "potentialPermitsTooltip": "Early information on your permitting needs will help us coordinate and expedite the authorization process.",
+    "projectApplicantTypeCard": "Are you applying as an individual or as a business?",
+    "projectDescriptionCard": "Help us learn more about your housing project",
     "projectLocationDescriptionCard": "Is there anything else you would like to tell us about this project's location? (optional)",
+    "projectNameCard": "Tell us your project name",
     "provincialPermitsCard": "Have you applied for any provincial permits for this project?",
-    "investigatePermitsCard": "Select all provincially issued permits you think you might need (optional)"
+    "rentalUnitsTooltip": "Rental refers to a purpose built residentual unit, property, or dwelling that is available for long term rent by tenants.",
+    "singleFamilySelectedCard": "Select the types of residential units being developed",
+    "validationBanner": "One or more of the required fields are missing or contains invalid data. Please check the highlighted pages and fields in red."
   },
   "submissionView": {
     "seePropViewButtonLabel": "See Proponent's View"
